+++ demos/components-off-canvas.php
@@ -57,16 +57,16 @@ $bp: (
               <ul>
                 <li><a href="">How It Works</a></li>
                 <li data-has-submenu>
+                  <input type="checkbox" id="subnav' . $nav . '-1" class="hidden-checkbox">
                   <label for="subnav' . $nav . '-1"><img src="images/arrow-r-w.png" alt=""></label>
                   <a href="">Compare</a>
-                  <input type="checkbox" id="subnav' . $nav . '-1" class="hidden-checkbox">
                   <ul data-submenu>
                     <li><label for="subnav' . $nav . '-1" data-back>Back</label></li>
                     <li><a href="">compare mode 1</a></li>
                     <li data-has-submenu>
+                      <input type="checkbox" id="subnav' . $nav . '-1-1" class="hidden-checkbox">
                       <label for="subnav' . $nav . '-1-1"><img src="images/arrow-r-w.png" alt=""></label>
                       <a href="">compare mode 2</a>
-                      <input type="checkbox" id="subnav' . $nav . '-1-1" class="hidden-checkbox">
                       <ul data-submenu>
                         <li><label for="subnav' . $nav . '-1-1" data-back>Back</label></li>
                         <li><a href="">item 1</a></li>
@@ -79,9 +79,9 @@ $bp: (
                 </li>
                 <li><a href="">Technology</a></li>
                 <li data-has-submenu>
+                  <input type="checkbox" id="subnav' . $nav . '-2" class="hidden-checkbox">
                   <label for="subnav' . $nav . '-2"><img src="images/arrow-r-w.png" alt=""></label>
                   <a href="">Careers</a>
-                  <input type="checkbox" id="subnav' . $nav . '-2" class="hidden-checkbox">
                   <ul data-submenu>
                     <li><label for="subnav' . $nav . '-2" data-back>Back</label></li>
                     <li><a href="">Accounting Manager</a></li>
@@ -122,16 +122,16 @@ $bp: (
           <ul>
             <li><a href="">How It Works</a></li>
             <li data-has-submenu>
+              <input type="checkbox" id="subnav' . $nav2 . '-1" class="hidden-checkbox">
               <label for="subnav' . $nav2 . '-1"><img src="images/arrow-r-w.png" alt=""></label>
               <a href="">Compare</a>
-              <input type="checkbox" id="subnav' . $nav2 . '-1" class="hidden-checkbox">
               <ul data-submenu>
                 <li><label for="subnav' . $nav2 . '-1" data-back>Back</label></li>
                 <li><a href="">compare mode 1</a></li>
                 <li data-has-submenu>
+                  <input type="checkbox" id="subnav' . $nav2 . '-1-1" class="hidden-checkbox">
                   <label for="subnav' . $nav2 . '-1-1"><img src="images/arrow-r-w.png" alt=""></label>
                   <a href="">compare mode 2</a>
-                  <input type="checkbox" id="subnav' . $nav2 . '-1-1" class="hidden-checkbox">
                   <ul data-submenu>
                     <li><label for="subnav' . $nav2 . '-1-1" data-back>Back</label></li>
                     <li><a href="">item 1</a></li>
@@ -144,9 +144,9 @@ $bp: (
             </li>
             <li><a href="">Technology</a></li>
             <li data-has-submenu>
+              <input type="checkbox" id="subnav' . $nav2 . '-2" class="hidden-checkbox">
               <label for="subnav' . $nav2 . '-2"><img src="images/arrow-r-w.png" alt=""></label>
               <a href="">Careers</a>
-              <input type="checkbox" id="subnav' . $nav2 . '-2" class="hidden-checkbox">
               <ul data-submenu>
                 <li><label for="subnav' . $nav2 . '-2" data-back>Back</label></li>
                 <li><a href="">Accounting Manager</a></li>