+++ demos/include/more-demos.php
@@ -3,7 +3,7 @@
   <h4>layout</h4>
   <nav class="demo-links">
     <?php 
-      $layouts = array('container', 'grid', 'gallery', 'metro', 'liquid-2', 'liquid-3', 'center', 'justify');
+      $layouts = array('container', 'grid', 'gallery', 'metro', 'liquid-2', 'liquid-3', 'center', 'justify', 'sticky-footer');
       $cur = '';
 
       foreach ($layouts as $layout) {