+++ demos/layout-center.php
@@ -7,16 +7,13 @@
       <h2 class="main-heading"><span>layout: </span>center</h2>
       <div class="example example-center" data-margin>
         <div data-content>
-        <div class="example-content">
-          <strong>Center align content</strong>
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Minus doloribus dicta voluptate quaerat vero!
-        </div>
+          <h4>Center align content</h4>
+          <img src="http://placehold.it/600x200/3DC3FF/3DC3FF" alt="">
+          <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Minus doloribus dicta voluptate quaerat vero!</p>
         </div>
       </div>
       <pre><code class="language-scss">
-.example-center { @include center(div left);
-  > div { width: 200px; padding: 10px; }
-}
+.example-center { @include center(div left); }
       </code></pre>
     </div>