+++ demos/layout-gallery.php
@@ -15,6 +15,15 @@
           <li><div class="cell" data-content></div></li>
           <li><div class="cell" data-content></div></li>
           <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
+          <li><div class="cell" data-content></div></li>
         </ul>
       </div>
       <pre><code class="language-scss">